- removed unused assets - created extension file - removed extra white space
