+++ recipe-app/View/CustomCollectionViewCell.swift
@@ -48,5 +48,3 @@ extension CustomCollectionViewCell {
         self.setUp()
     }
 }
-
-
