+++ recipe-app/ViewModel/RecipeListViewModel.swift
@@ -8,8 +8,8 @@
 import UIKit
 
 final class RecipeListViewModel {
-     var recipes = [Recipe]()
-     var images: [String] = []
+    var recipes = [Recipe]()
+    var images: [String] = []
     
     func searchRecipe(_ query: String,
                       completion: @escaping (Result<[Recipe], Error>) -> Void)  {
@@ -59,22 +59,3 @@ extension RecipeListViewModel {
         return components?.url
     }
 }
-
-extension UIImageView {
-    public func setImage(from urlString: String, placeholder: UIImage? = nil) {
-        guard let url = URL(string: urlString) else {
-            self.image = placeholder
-            return
-        }
-        
-        self.image = placeholder
-        let task = URLSession.shared.dataTask(with: url) { [weak self] (data, response, error) in
-            guard let imageData = data, let image = UIImage(data: imageData) else { return }
-            DispatchQueue.main.async {
-                self?.image = image
-            }
-        }
-        task.resume()
-    }
-}
-
